+++ tests/integration/bff-express/src/modern-app-env.d.ts
@@ -2,4 +2,3 @@
 /// <reference types='@modern-js/plugin-testing/types' />
 /// <reference types='@modern-js/plugin-micro-frontend/type' />
 /// <reference types='@modern-js/plugin-express/types' />
-/// <reference types='@modern-js/plugin-koa/types' />