+++ packages/runtime/plugin-testing/CHANGELOG.md
@@ -1,5 +1,24 @@
 # @modern-js/plugin-testing
 
+## 2.34.0
+
+### Patch Changes
+
+- Updated dependencies [f851fa9]
+- Updated dependencies [7d70738]
+- Updated dependencies [5240e5d]
+- Updated dependencies [a77b82a]
+- Updated dependencies [dcdeea5]
+- Updated dependencies [eca24a2]
+- Updated dependencies [c8b448b]
+  - @modern-js/prod-server@2.34.0
+  - @modern-js/utils@2.34.0
+  - @modern-js/runtime@2.34.0
+  - @modern-js/babel-preset-app@2.34.0
+  - @modern-js/babel-compiler@2.34.0
+  - @modern-js/plugin@2.34.0
+  - @modern-js/types@2.34.0
+
 ## 2.33.1
 
 ### Patch Changes